add illumina.py flowcell_metadata command
add illumina.py flowcell_metadata command to write sequencing machine name and lane cont based on either a run directory, a RunInfo.xml file, or a flowcell ID string. This relies on a new function in the RunInfo() class, infer_sequencer_model(), that tries to identify a machine based on the suffix of its flowcell ID, or failing that, the tile count (requiring a RunInfo.xml file). This also adds to the TestRunInfo() unit tests to check the machine type of the test RunInfo files