in util.version.py::get_project_path(), consider calling modules
When derived modules (i.e. viral-classify) are symlinked into viral-core and their versions are read from a VERSION file, only the path to the viral-core version file was used. This changes that behavior, so outer members of the call stack are checked for the presence of VERSION files as well (and if found, the version is read from the first/outermost one found).

(i.e. when calling ./taxon-filter.py --version, with these changes, the version of viral-classify is correctly returned rather than the version of viral-core)

Since this relies on the presence of a VERSION file to find the current "project path", this will fall back to the old behavior if viral-core has a VERSION file and a derived module lacks such a file.

This does not change the behavior where a version file will not be created for a derived module when calling `./symlinked_derived_module.py --version`